+++ Character_Management.Application/Features/Characters/Handlers/Commands/DeleteCharacterCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using Character_Management.Application.Exceptions;
 using Character_Management.Application.Features.Characters.Requests.Commands;
 using Character_Management.Application.persistance.contracts;
+using Character_Management.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -24,6 +26,10 @@ namespace Character_Management.Application.Features.Characters.Handlers.Commands
         public async Task<Unit> Handle(DeleteCharacterCommand request, CancellationToken cancellationToken)
         {
             var character = await _characterRepository.Get(request.ID);
+            if(character == null)
+            {
+                throw new NotFoundException(nameof(Character), request.ID);
+            }
             await _characterRepository.Delete(character);
             return Unit.Value;
         }
